+++ Base/controllers/UsersController.php
@@ -2,7 +2,7 @@
 class UsersController {
 
     public function listAdmin() {
-        checkRole(['admin']);
+        requireAdmin();
         $user = new UserModel();
         $listUser = $user->getList();
 
@@ -10,7 +10,7 @@ class UsersController {
         require_once PATH_VIEW . "main.php";
     }
     public function listCustomer() {
-        checkRole(['admin']);
+        requireAdmin();
         $user = new UserModel();
         $listUser = $user->getList();
 
@@ -19,7 +19,7 @@ class UsersController {
     }
 
     public function createUser() {
-        checkRole(['admin']);
+        requireAdmin();
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
 
             $model = new UserModel();
@@ -40,7 +40,7 @@ class UsersController {
     }
 
     public function updateUser() {
-        checkRole(['admin']);
+        requireAdmin();
         $model = new UserModel();
 
         if ($_SERVER['REQUEST_METHOD'] === 'GET') {
@@ -74,7 +74,7 @@ class UsersController {
     }
 
     public function deleteUser() {
-        checkRole(['admin']);
+        requireAdmin();
         $model = new UserModel();
         $model->delete($_GET['id']);
         header("Location: " . BASE_URL . "?action=list-customer");
